requirements.txt
# LangGraph 和 LLM 相关
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# Neo4j（使用稳定版本避免 pandas/numpy 冲突）
neo4j==5.14.0

# PDF 处理
PyMuPDF>=1.24.0
pdfplumber>=0.11.0

# HTTP 客户端
httpx>=0.27.0

# 工具库
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 日志
loguru>=0.7.0

# 兼容性修复
numpy<2.0.0
